pyproject.toml
[project]
name = "text-to-sql-eval-lab"
version = "0.1.0"
description = "LLM evaluation and optimization lab — Text-to-SQL use case"
requires-python = ">=3.11"
dependencies = [
    "inspect-ai>=0.3.0",
    "litellm>=1.40.0",
    "langfuse>=2.0.0",
    "sqlglot>=23.0.0",      # SQL parsing and validation
    "duckdb>=0.10.0",        # Fast in-process SQL engine (replaces SQLite for analytics queries)
    "rich>=13.0.0",          # Pretty terminal output
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "dspy-ai>=2.4.5",
    "chromadb>=1.0.0",
    "google-genai",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ipykernel>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
